requirements.txt
python==3.7
astroid==2.4.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
decorators==2.0.7
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1
django-admin-interface==0.16.3
django-admin-tools==0.9.1
django-cleanup==5.1.0
django-colorfield==0.4.1
django-crispy-forms==1.9.2
django-embed-video==1.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-import-export==2.5.0
django-object-actions==3.0.2
django-shopping-cart==0.1
django-smtp-ssl==1.0
Django-Verify-Email==0.0.5
et-xmlfile==1.1.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==4.0.1
isort==4.3.21
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
numpy==1.20.3
odfpy==1.4.1
openpyxl==3.0.7
Pillow==7.1.2
polls==0.1
psycopg2==2.8.6
pylint==2.5.2
pyrsistent==0.17.3
pystache==0.5.4
python-http-client==3.3.2
pytz==2020.1
PyYAML==5.4.1
ranges==0.3
razorpay==1.2.0
requests==2.25.1
sendgrid==6.6.0
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
static==1.1.1
static-ranges==0.2.0
static3==0.7.0
tablib==3.0.0
toml==0.10.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
whitenoise==5.0.1
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
